Update transaction value to numeric

+++ borgs-db/create.sql
@@ -43,7 +43,7 @@ CREATE TABLE Transactions (
 	transaction_id SERIAL PRIMARY KEY,
 	virtual_account INT NOT NULL REFERENCES VirtualAccounts(account_id),
 	physical_account INT NOT NULL REFERENCES PhysicalAccounts(account_id),
-	value REAL NOT NULL,
+	value NUMERIC NOT NULL,
 	category INT NOT NULL REFERENCES TransactionsCategories(category_id),
 	timestampepochseconds INT NOT NULL, -- Date stored in unix/epoch time
 	description TEXT,
